+++ test/inputs/collection_check_boxes_input_test.rb
@@ -300,4 +300,11 @@ class CollectionCheckBoxesInputTest < ActionView::TestCase
       assert_select 'label[for=user_1_gender_female]'
     end
   end
+
+  test 'collection input with check_boxes type and item_extra_html is included' do
+    with_input_for @user, :name, :check_boxes, collection: [['Jose', 1], ['Carlos', 2]],
+                          item_extra_html: -> (item, value) { { 'data-extra' => "#{item.first}-and-#{value}" } }
+    assert_select 'input[value="1"][data-extra="Jose-and-1"]'
+    assert_select 'input[value="2"][data-extra="Carlos-and-2"]'
+  end
 end